third_party/WebKit/LayoutTests/fast/dom/shadow/disconnected-distribution-cycle.html

<!DOCTYPE html>

<script src="../../../resources/js-test.js"></script>

<div id="b"></div>

<script>
    description("Distributions from disconnected subtrees should be cleared when inserted again to avoid cycles.");

    var a = document.createElement("div");
    var b = document.getElementById("b");

    // Put "a" it into a disconnected element with a shadow root with an insertion point.
    var host = document.createElement("div");
    var content = host.createShadowRoot().appendChild(document.createElement("content"));
    host.appendChild(a);

    // Cause a distribution of "a" into the disconnected host.
    a.getDestinationInsertionPoints();

    // Distribute "a" into a host in the page.
    b.createShadowRoot().appendChild(document.createElement("content"));
    b.appendChild(a);

    // Cause an attach of all elements in the page.
    getComputedStyle(document.body).color;

    // Insert the disconnected host as a child of "a".
    a.appendChild(host);

    // Go through distribution again... CRASH!
    shouldBe("content.getDistributedNodes().length", "0");
</script>
