third_party/WebKit/LayoutTests/fast/dom/shadow/disconnected-distribution-cycle-expected.txt

Distributions from disconnected subtrees should be cleared when inserted again to avoid cycles.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS content.getDistributedNodes().length is 0
PASS successfullyParsed is true

TEST COMPLETE
